Extract fields from state via extract_op_{scalar,vector,tensor}_field
The Fluidity OP2 interface provides the functions
extract_op_{scalar,vector,tensor}_field to extract a field from state,
returning a struct containing the dat and map associated to that field
for the requested co-dimension (for a dx integral we want cells i.e.
co-dimension 0).

Note: We always request co-dimension 0 for now.